+++ src/airfoil_aero/AirfoilAero.jl
@@ -1,5 +1,6 @@
 module AirfoilAero
 
+using Arrow
 using LinearAlgebra
 using Parameters
 using Statistics